+++ api/src/main/java/io/strimzi/api/kafka/model/kafka/KafkaClusterSpec.java
@@ -22,6 +22,7 @@ import io.strimzi.api.kafka.model.common.jmx.HasJmxOptions;
 import io.strimzi.api.kafka.model.common.jmx.KafkaJmxOptions;
 import io.strimzi.api.kafka.model.common.metrics.MetricsConfig;
 import io.strimzi.api.kafka.model.kafka.listener.GenericKafkaListener;
+import io.strimzi.api.kafka.model.kafka.tieredstorage.TieredStorage;
 import io.strimzi.crdgenerator.annotations.AddedIn;
 import io.strimzi.crdgenerator.annotations.Description;
 import io.strimzi.crdgenerator.annotations.DescriptionFile;
@@ -39,7 +40,7 @@ import java.util.Map;
 /**
  * Representation of a Strimzi-managed Kafka "cluster".
  */
-@DescriptionFile 
+@DescriptionFile
 @Buildable(
         editableEnabled = false,
         builderPackage = Constants.FABRIC8_KUBERNETES_API
@@ -84,6 +85,7 @@ public class KafkaClusterSpec implements HasConfigurableMetrics, HasConfigurable
     private List<GenericKafkaListener> listeners;
     private KafkaAuthorization authorization;
     private KafkaClusterTemplate template;
+    private TieredStorage tieredStorage;
     private Map<String, Object> additionalProperties = new HashMap<>(0);
 
     @Description("The Kafka broker version. Defaults to the latest version. " +
@@ -173,7 +175,7 @@ public class KafkaClusterSpec implements HasConfigurableMetrics, HasConfigurable
     }
 
     @Description("The container image used for Kafka pods. "
-        + "If the property is not set, the default Kafka image version is determined based on the `version` configuration. "  
+        + "If the property is not set, the default Kafka image version is determined based on the `version` configuration. "
         + "The image names are specifically mapped to corresponding versions in the Cluster Operator configuration. "
         + "Changing the Kafka image version does not automatically update the image versions for other components, such as Kafka Exporter. ")
     @JsonInclude(JsonInclude.Include.NON_DEFAULT)
@@ -279,6 +281,16 @@ public class KafkaClusterSpec implements HasConfigurableMetrics, HasConfigurable
 
     public void setTemplate(KafkaClusterTemplate template) {
         this.template = template;
+    }
+
+    @Description("Configure the tiered storage feature for Kafka brokers")
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    public TieredStorage getTieredStorage() {
+        return tieredStorage;
+    }
+
+    public void setTieredStorage(TieredStorage tieredStorage) {
+        this.tieredStorage = tieredStorage;
     }
 
     @Override